split AddressAutocompleteInput into directory
